requirements.txt
anyio==3.6.2
asgiref==3.4.1
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.7.1
boto3==1.23.10
botocore==1.26.10
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==8.0.4
conda==4.6.14
conda-build==3.17.8
contextlib2==21.6.0
contextvars==2.4
cryptography==2.6.1
cupy-cuda100==9.6.0
dataclasses==0.8
decorator==4.4.0
fastapi==0.83.0
fastrlock==0.8.2
filelock==3.0.10
gdown==4.7.3
glob2==0.6
h11==0.13.0
idna==2.8
immutables==0.19
importlib-metadata==4.8.3
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.10.0
libarchive-c==2.8
lief==0.9.0
MarkupSafe==1.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
numpy==1.19.5
olefile==0.46
opencv-contrib-python==3.4.1.15
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkginfo==1.5.0.1
prompt-toolkit==2.0.9
psutil==5.6.2
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
pydantic==1.9.2
Pygments==2.3.1
pyOpenSSL==19.0.0
PySocks==1.6.8
python-dateutil==2.9.0.post0
python-multipart==0.0.5
pytz==2019.1
PyYAML==5.1
requests==2.21.0
ruamel-yaml==0.15.46
s3transfer==0.5.2
six==1.12.0
sniffio==1.2.0
soupsieve==1.8
starlette==0.19.1
torch==1.1.0
torchvision==0.2.2
tqdm==4.31.1
traitlets==4.3.2
typing-extensions==4.1.1
urllib3==1.26.18
uvicorn==0.16.0
wcwidth==0.1.7
zipp==3.6.0